fix(core): Simplify Git branch listing and handle deleted branches (no-changelog)
