+++ packages/cli/src/modules/n8n-packages/entities/folder/__tests__/folder.exporter.test.ts
@@ -130,25 +130,25 @@ describe('FolderExporter', () => {
 	});
 
 	it('writes only the folders on the path to a selected workflow, keeping sibling slugs', async () => {
-		const opsA = makeFolder({ id: 'ops-a', name: 'Ops', createdAt: new Date('2026-01-01') });
-		const opsB = makeFolder({ id: 'ops-b', name: 'Ops', createdAt: new Date('2026-02-01') });
-		const nested = makeFolder({ id: 'nested', name: 'Nested', parentFolderId: 'ops-b' });
+		const opsA = makeFolder({ id: 'ops_a', name: 'Ops', createdAt: new Date('2026-01-01') });
+		const opsB = makeFolder({ id: 'ops_b', name: 'Ops', createdAt: new Date('2026-02-01') });
+		const nested = makeFolder({ id: 'nested', name: 'Nested', parentFolderId: 'ops_b' });
 		const { exporter, workflowFinder, workflowExporter } = makeExporter([opsA, opsB, nested]);
 		workflowFinder.findWorkflowIdsByFolder.mockResolvedValue(
 			new Map([
-				['ops-a', ['w-a']],
+				['ops_a', ['w-a']],
 				['nested', ['w-n1', 'w-n2']],
 			]),
 		);
 		workflowExporter.export.mockResolvedValue({
-			entries: [{ id: 'w-n2', name: 'N2', target: 'folders/ops-ops-b/nested-nested/workflows/n2' }],
+			entries: [{ id: 'w-n2', name: 'N2', target: 'folders/ops-ops_b/nested-nested/workflows/n2' }],
 			requirements: { credentials: [], dataTables: [], variables: [], tags: [], nodeTypes: [] },
 		});
 		const writer = new CapturingWriter();
 
 		const result = await exporter.export({
 			user,
-			folderIds: ['ops-a', 'ops-b'],
+			folderIds: ['ops_a', 'ops_b'],
 			selectedWorkflowIds: new Set(['w-n2']),
 			writer,
 			includeTags: true,
@@ -157,16 +157,16 @@ describe('FolderExporter', () => {
 		});
 
 		expect(result.entries.map((e) => e.target)).toEqual([
-			'folders/ops-ops-b',
-			'folders/ops-ops-b/nested-nested',
+			'folders/ops-ops_b',
+			'folders/ops-ops_b/nested-nested',
 		]);
-		expect(writer.directories).toEqual(['folders/ops-ops-b', 'folders/ops-ops-b/nested-nested']);
+		expect(writer.directories).toEqual(['folders/ops-ops_b', 'folders/ops-ops_b/nested-nested']);
 		// The unselected sibling never reaches the workflow exporter, so it is never fetched.
 		expect(workflowExporter.export).toHaveBeenCalledTimes(1);
 		expect(workflowExporter.export).toHaveBeenCalledWith(
 			expect.objectContaining({
 				workflowIds: ['w-n2'],
-				basePrefix: 'folders/ops-ops-b/nested-nested',
+				basePrefix: 'folders/ops-ops_b/nested-nested',
 			}),
 		);
 		expect(result.workflowEntries.map((e) => e.id)).toEqual(['w-n2']);
