+++ packages/cli/src/modules/promotions.ee/__tests__/promotions-package.integration.test.ts
@@ -2,24 +2,36 @@ import { LicenseState } from '@n8n/backend-common';
 import {
 	createTeamProject,
 	createWorkflow,
+	getPersonalProject,
 	mockInstance,
 	mockLogger,
 	testDb,
 	testModules,
 } from '@n8n/backend-test-utils';
-import type { User } from '@n8n/db';
+import type { Project, User } from '@n8n/db';
 import { FolderRepository, ProjectRepository, WorkflowRepository } from '@n8n/db';
 import { Container } from '@n8n/di';
 import type { Cipher, InstanceSettings } from 'n8n-core';
 import { jsonParse } from 'n8n-workflow';
 import assert from 'node:assert';
-import { mkdtemp, readFile, rm, writeFile } from 'node:fs/promises';
+import { mkdir, mkdtemp, readdir, readFile, rename, rm, writeFile } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import path from 'node:path';
 import { simpleGit, type SimpleGit } from 'simple-git';
 import { mock } from 'vitest-mock-extended';
 
 import { ActiveWorkflowManager } from '@/active-workflow-manager';
+import { BadRequestError } from '@/errors/response-errors/bad-request.error';
+import { mockDataTableSizeValidator } from '@/modules/data-table/__tests__/test-helpers';
+import { DataTableService } from '@/modules/data-table/data-table.service';
+import {
+	PACKAGE_ENTITY_LAYOUT,
+	entityFilePath,
+	workflowLifecycleFilePath,
+	type ManifestEntityCollection,
+} from '@/modules/n8n-packages/io/manifest-entry';
+import { saveCredential } from '@test-integration/db/credentials';
+import { createTag } from '@test-integration/db/tags';
 import { VariablesService } from '@/environments.ee/variables/variables.service.ee';
 import { buildWorkflowReferencingVariables } from '@/modules/n8n-packages/__tests__/utils/test-builders';
 import { N8nPackagesService } from '@/modules/n8n-packages/n8n-packages.service';
@@ -64,9 +76,10 @@ let packagesService: N8nPackagesService;
 let owner: User;
 let testRoot: string;
 let service: PromotionsService;
+let workingDirectory: PromotionWorkingDirectoryService;
 
 beforeAll(async () => {
-	await testModules.loadModules(['n8n-packages', 'promotions']);
+	await testModules.loadModules(['n8n-packages', 'promotions', 'data-table']);
 	await testDb.init();
 
 	providerRepository = Container.get(PromotionProviderRepository);
@@ -95,6 +108,12 @@ beforeEach(async () => {
 	await connectionRepository.delete({});
 	await providerRepository.delete({});
 	await testDb.truncate([
+		'WorkflowTagMapping',
+		'TagEntity',
+		'CredentialsEntity',
+		'SharedCredentials',
+		'DataTable',
+		'DataTableColumn',
 		'Folder',
 		'WorkflowEntity',
 		'SharedWorkflow',
@@ -103,6 +122,7 @@ beforeEach(async () => {
 		'Variables',
 	]);
 	await Container.get(VariablesService).updateCache();
+	mockDataTableSizeValidator();
 	licenseMocker.reset();
 	owner = await createOwner();
 	testRoot = await mkdtemp(path.join(tmpdir(), 'n8n-promotions-roundtrip-'));
@@ -113,6 +133,9 @@ beforeEach(async () => {
 	cipher.encryptV2.mockImplementation(async (value) => String(value));
 	const logger = mockLogger();
 	const gitService = new PromotionsGitService(logger);
+	workingDirectory = new PromotionWorkingDirectoryService(
+		mock<InstanceSettings>({ n8nFolder: path.join(testRoot, 'instance') }),
+	);
 	service = new PromotionsService(
 		new PromotionConfigResolver(
 			configRepository,
@@ -121,9 +144,7 @@ beforeEach(async () => {
 			projectRepository,
 		),
 		new PromotionProvidersService(providerRepository, connectionRepository, gitService, cipher),
-		new PromotionWorkingDirectoryService(
-			mock<InstanceSettings>({ n8nFolder: path.join(testRoot, 'instance') }),
-		),
+		workingDirectory,
 		gitService,
 		projectRepository,
 		projectService,
@@ -205,6 +226,36 @@ async function createInstanceConnection(
 		},
 	});
 	return connection;
+}
+
+async function writeRemoteFile(remote: TestRemote, relativePath: string, content: string) {
+	const filePath = path.join(remote.workingDir, relativePath);
+	await mkdir(path.dirname(filePath), { recursive: true });
+	await writeFile(filePath, content);
+}
+
+async function commitAndPushRemote(remote: TestRemote, message: string) {
+	await remote.git.add(['--all']);
+	await remote.git.commit(message);
+	await remote.git.push('origin', 'main');
+}
+
+async function remoteBlobSha(remote: TestRemote, filePath: string): Promise<string> {
+	return (await simpleGit(remote.bareDir).raw(['rev-parse', `main:${filePath}`])).trim();
+}
+
+async function snapshotWorkingTree(dir: string): Promise<Map<string, string>> {
+	const snapshot = new Map<string, string>();
+	const walk = async (current: string) => {
+		for (const entry of await readdir(current, { withFileTypes: true })) {
+			if (entry.name === '.git') continue;
+			const entryPath = path.join(current, entry.name);
+			if (entry.isDirectory()) await walk(entryPath);
+			else snapshot.set(path.relative(dir, entryPath), await readFile(entryPath, 'utf-8'));
+		}
+	};
+	await walk(dir);
+	return snapshot;
 }
 
 describe('Promote and Apply', () => {
@@ -437,5 +488,303 @@ describe('Promote and Apply', () => {
 		await expect(
 			readFile(path.join(inspectionDir, 'n8n-export', 'manifest.json')),
 		).rejects.toThrow();
+	});
+});
+
+describe('Promotion base branch listing', () => {
+	let project: Project;
+
+	beforeEach(async () => {
+		project = await createTeamProject('Orders', owner);
+	});
+
+	it('lists the exported project and its top-level dependencies', async () => {
+		const remote = await createRemote();
+		const connection = await createInstanceConnection(remote.bareDir);
+		await service.clone(connection.id, 'promote');
+
+		const projectB = await createTeamProject('Marketing', owner);
+		const parentFolder = await createFolder(project, { name: 'Operations' });
+		const childFolder = await createFolder(project, { name: 'Orders', parentFolder });
+		const dataTableService = Container.get(DataTableService);
+		const projectTable = await dataTableService.createDataTable(project.id, {
+			name: 'Orders',
+			columns: [{ name: 'email', type: 'string' }],
+		});
+		const sharedTable = await dataTableService.createDataTable(
+			(await getPersonalProject(owner)).id,
+			{
+				name: 'Customers',
+				columns: [{ name: 'email', type: 'string' }],
+			},
+		);
+		const credential = await saveCredential(
+			{
+				name: 'Header credential',
+				type: 'httpHeaderAuth',
+				data: { name: 'X-Auth', value: 'secret' },
+			},
+			{ user: owner, role: 'credential:owner' },
+		);
+		await createVariable('API_URL', 'https://api.example.com');
+		const workflow = await createWorkflow(
+			{
+				name: 'Process order',
+				parentFolder: childFolder,
+				nodes: [
+					{
+						id: 'n1',
+						name: 'HTTP',
+						type: 'n8n-nodes-base.httpRequest',
+						typeVersion: 1,
+						position: [0, 0],
+						parameters: { url: '={{ $vars.API_URL }}' },
+						credentials: { httpHeaderAuth: { id: credential.id, name: credential.name } },
+					},
+					...[projectTable, sharedTable].map((table, index) => ({
+						id: `table${index}`,
+						name: table.name,
+						type: 'n8n-nodes-base.dataTable',
+						typeVersion: 1,
+						position: [index * 200, 200] as [number, number],
+						parameters: { dataTableId: { __rl: true, mode: 'id', value: table.id } },
+					})),
+				],
+				connections: {},
+			},
+			project,
+		);
+		const tag = await createTag({ name: 'prod' }, workflow);
+		await createWorkflow({ name: 'Campaign', nodes: [], connections: {} }, projectB);
+
+		await service.promote(connection.id, owner, {
+			canExportVariableValues: true,
+			commitMessage: 'Export projects',
+		});
+		const manifest = packageManifestSchema.parse(
+			jsonParse(await simpleGit(remote.bareDir).show(['main:n8n-export/manifest.json'])),
+		);
+		const projectEntry = manifest.projects?.find(({ id }) => id === project.id);
+		assert(projectEntry);
+		const collections = Object.keys(PACKAGE_ENTITY_LAYOUT) as ManifestEntityCollection[];
+		const expectedPaths = collections.flatMap((collection) =>
+			(manifest[collection] ?? [])
+				.filter(
+					({ target }) =>
+						!target.startsWith(`${PACKAGE_ENTITY_LAYOUT.projects.directory}/`) ||
+						target === projectEntry.target ||
+						target.startsWith(`${projectEntry.target}/`),
+				)
+				.flatMap(({ target }) => [
+					`n8n-export/${entityFilePath(collection, target)}`,
+					...(collection === 'workflows'
+						? [`n8n-export/${workflowLifecycleFilePath(target)}`]
+						: []),
+				]),
+		);
+
+		const files = await service.listBaseBranchFiles(project.id);
+
+		expect(files.map(({ key, type }) => ({ key, type }))).toEqual(
+			expect.arrayContaining([
+				{ key: project.id, type: 'project' },
+				{ key: parentFolder.id, type: 'folder' },
+				{ key: childFolder.id, type: 'folder' },
+				{ key: workflow.id, type: 'workflow' },
+				{ key: credential.id, type: 'credential' },
+				{ key: 'apiurl', type: 'variable' },
+				{ key: tag.id, type: 'tag' },
+				{ key: projectTable.id, type: 'dataTable' },
+				{ key: sharedTable.id, type: 'dataTable' },
+			]),
+		);
+		expect(files.map(({ path }) => path).sort()).toEqual(expectedPaths.sort());
+
+		for (const file of files) {
+			expect(file.blobSha).toBe(await remoteBlobSha(remote, file.path));
+		}
+	});
+
+	it('uses the project promote config instead of the instance or apply config', async () => {
+		const remote = await createRemote();
+		const projectPath = `n8n-export/projects/orders-${project.id}/project.json`;
+		await writeRemoteFile(remote, projectPath, '{"name":"Main"}');
+		await commitAndPushRemote(remote, 'Export main project');
+		await remote.git.checkoutLocalBranch('production');
+		await writeRemoteFile(remote, projectPath, '{"name":"Production"}');
+		await remote.git.add(['--all']);
+		await remote.git.commit('Export production project');
+		await remote.git.push('origin', 'production');
+		const instance = await createInstanceConnection(remote.bareDir, {
+			apply: 'production',
+			promote: 'main',
+		});
+		await service.clone(instance.id, 'promote');
+		expect(await service.listBaseBranchFiles(project.id)).toEqual([
+			{
+				key: project.id,
+				path: projectPath,
+				blobSha: await remoteBlobSha(remote, projectPath),
+				type: 'project',
+			},
+		]);
+
+		const connection = await connectionRepository.insertConnection({
+			name: 'Project production',
+			scope: 'projects',
+			providerId: instance.providerId,
+			target: { schemaVersion: 1, remoteUrl: remote.bareDir },
+		});
+		await configRepository.insertConfig({
+			connectionId: connection.id,
+			direction: 'promote',
+			name: 'Promote',
+			settings: { schemaVersion: 1, baseBranchName: 'production', createBranchOnPromotion: false },
+		});
+		await linkRepository.linkProject(project.id, connection.id);
+		await service.clone(connection.id, 'promote');
+
+		expect(await service.listBaseBranchFiles(project.id)).toEqual([
+			{
+				key: project.id,
+				path: projectPath,
+				blobSha: (await remote.git.revparse([`production:${projectPath}`])).trim(),
+				type: 'project',
+			},
+		]);
+	});
+
+	it('reads remote updates without changing a dirty, detached checkout', async () => {
+		const remote = await createRemote();
+		const projectPath = `n8n-export/projects/ünïcode örders-${project.id}/project.json`;
+		const variablePath = 'n8n-export/variables/my "quoted" var-Va45zz67/variable.json';
+		await writeRemoteFile(remote, projectPath, '{"name":"Ünïcode örders"}');
+		await writeRemoteFile(remote, variablePath, '{"name":"my var"}');
+		await commitAndPushRemote(remote, 'Initial export');
+
+		const connection = await createInstanceConnection(remote.bareDir);
+		await service.clone(connection.id, 'promote');
+
+		const config = await configRepository.findByConnectionAndDirection(connection.id, 'promote');
+		assert(config);
+		const checkout = workingDirectory.paths(config.id).repositoryFolder;
+		const checkoutGit = simpleGit(checkout);
+		await checkoutGit.addConfig('core.autocrlf', 'true');
+		await checkoutGit.raw(['checkout', '--detach']);
+		await writeFile(path.join(checkout, projectPath), 'Local changes\r\n');
+		await writeFile(path.join(checkout, 'untracked.txt'), 'Untracked file\n');
+		const headBefore = (await checkoutGit.revparse(['HEAD'])).trim();
+		const treeBefore = await snapshotWorkingTree(checkout);
+
+		const firstListing = await service.listBaseBranchFiles(project.id);
+		expect(firstListing).toHaveLength(2);
+		expect(firstListing).toContainEqual({
+			key: 'my "quoted" var',
+			path: variablePath,
+			blobSha: await remoteBlobSha(remote, variablePath),
+			type: 'variable',
+		});
+
+		const workflowPath = `n8n-export/projects/ünïcode örders-${project.id}/workflows/my-hyphen-ated-slug-Wf99zz88/workflow.json`;
+		await writeRemoteFile(remote, workflowPath, '{"name":"My hyphen-ated workflow"}');
+		await commitAndPushRemote(remote, 'Add workflow');
+
+		const secondListing = await service.listBaseBranchFiles(project.id);
+
+		expect(secondListing).toContainEqual({
+			key: 'Wf99zz88',
+			path: workflowPath,
+			blobSha: await remoteBlobSha(remote, workflowPath),
+			type: 'workflow',
+		});
+		expect((await checkoutGit.revparse(['HEAD'])).trim()).toBe(headBefore);
+		expect(await snapshotWorkingTree(checkout)).toEqual(treeBefore);
+	});
+
+	it('returns an empty listing when the remote branch has no commits yet', async () => {
+		const bareDir = path.join(testRoot, 'empty-remote.git');
+		await simpleGit().raw(['init', '--bare', bareDir]);
+		const connection = await createInstanceConnection(bareDir);
+		await service.clone(connection.id, 'promote');
+
+		const files = await service.listBaseBranchFiles(project.id);
+
+		expect(files).toEqual([]);
+	});
+
+	it('preserves files with matching IDs or variable slugs across collections and scopes', async () => {
+		const remote = await createRemote();
+		const projectRoot = `n8n-export/projects/orders-${project.id}`;
+		const entities = [
+			{ key: '42', type: 'folder', path: `${projectRoot}/folders/legacy-42/folder.json` },
+			{
+				key: '42',
+				type: 'workflow',
+				path: `${projectRoot}/folders/legacy-42/workflows/order-42/workflow.json`,
+			},
+			{ key: '42', type: 'credential', path: `${projectRoot}/credentials/api-42/credential.json` },
+			{ key: 'apiurl', type: 'variable', path: `${projectRoot}/variables/apiurl-1/variable.json` },
+			{ key: 'apiurl', type: 'variable', path: 'n8n-export/variables/apiurl-2/variable.json' },
+		];
+		for (const entity of entities) {
+			await writeRemoteFile(remote, entity.path, '{}');
+		}
+		await writeRemoteFile(remote, 'n8n-export/manifest.json', 'Not used for this listing');
+		await writeRemoteFile(remote, 'n8n-export/workflows/standalone-Wf01/workflow.json', '{}');
+		await writeRemoteFile(remote, 'n8n-export/folders/standalone-Fo01/folder.json', '{}');
+		await writeRemoteFile(remote, `${projectRoot}/workflows/order-42/README.md`, 'Notes');
+		await commitAndPushRemote(remote, 'Export scoped entities');
+		const connection = await createInstanceConnection(remote.bareDir);
+		await service.clone(connection.id, 'promote');
+
+		const files = await service.listBaseBranchFiles(project.id);
+
+		expect(files.map(({ blobSha, ...entity }) => entity)).toEqual(expect.arrayContaining(entities));
+		expect(files).toHaveLength(entities.length);
+	});
+
+	it('reports an unavailable remote before its first commit', async () => {
+		const bareDir = path.join(testRoot, 'empty-remote.git');
+		await simpleGit().raw(['init', '--bare', bareDir]);
+		const connection = await createInstanceConnection(bareDir);
+		await service.clone(connection.id, 'promote');
+		await rename(bareDir, `${bareDir}.offline`);
+
+		await expect(service.listBaseBranchFiles(project.id)).rejects.toThrow(BadRequestError);
+	});
+
+	it('requires a fresh clone after a base-branch change and reports a deleted branch', async () => {
+		const remote = await createRemote();
+		const connection = await createInstanceConnection(remote.bareDir);
+		await service.clone(connection.id, 'promote');
+		await remote.git.checkoutLocalBranch('production');
+		const projectPath = `n8n-export/projects/orders-${project.id}/project.json`;
+		await writeRemoteFile(remote, projectPath, '{}');
+		await remote.git.add(['--all']);
+		await remote.git.commit('Export production project');
+		await remote.git.push('origin', 'production');
+		const config = await configRepository.findByConnectionAndDirection(connection.id, 'promote');
+		assert(config);
+		await configRepository.replaceConfig(config.id, {
+			name: 'Promote',
+			settings: { schemaVersion: 1, baseBranchName: 'production', createBranchOnPromotion: false },
+		});
+		await expect(service.listBaseBranchFiles(project.id)).rejects.toThrow('not cloned');
+		await service.clone(connection.id, 'promote');
+
+		const files = await service.listBaseBranchFiles(project.id);
+
+		expect(files).toEqual([
+			{
+				key: project.id,
+				path: projectPath,
+				blobSha: (await remote.git.revparse([`production:${projectPath}`])).trim(),
+				type: 'project',
+			},
+		]);
+
+		await remote.git.raw(['push', 'origin', '--delete', 'production']);
+		await simpleGit(remote.bareDir).raw(['update-ref', '-d', 'refs/heads/main']);
+		await expect(service.listBaseBranchFiles(project.id)).rejects.toThrow(BadRequestError);
 	});
 });
